+++ src/resolvers/query/index.js
@@ -1,4 +1,5 @@
 import paymentQueries from './payment';
+import RequestQueries from './request';
 import subscriptionQueries from './subscription';
 import UserQueries from './user';
 
@@ -7,6 +8,7 @@ const Query = {
   ...UserQueries,
   ...paymentQueries,
   ...subscriptionQueries,
+  ...RequestQueries,
 };
 
 export default Query;